feat: Add sized array types
