+++ modules/core/src/classes/matrix3.ts
@@ -3,7 +3,7 @@
 // Copyright (c) vis.gl contributors
 // Copyright (c) 2017 Uber Technologies, Inc.
 
-import {NumericArray} from '@math.gl/types';
+import {NumericArray, NumberArray9} from '@math.gl/types';
 import {Matrix} from './base/matrix';
 import {checkVector} from '../lib/validators';
 
@@ -36,6 +36,13 @@ enum INDICES {
 
 const IDENTITY_MATRIX = Object.freeze([1, 0, 0, 0, 1, 0, 0, 0, 1]);
 
+/** Helper type that captures array length for a 3x3 matrix */
+export type Matrix3Like = Matrix3 | NumberArray9;
+
+/** 
+ * A 3x3 matrix with common linear algebra operations
+ * Subclass of Array<number> meaning that it is highly compatible with other libraries
+ */
 export class Matrix3 extends Matrix {
   static get IDENTITY(): Readonly<Matrix3> {
     return getIdentityMatrix();
