+++ modules/types/src/index.ts
@@ -2,5 +2,18 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
-export type {TypedArray, TypedArrayConstructor, NumericArray, NumberArray} from './array-types';
+export type {
+  TypedArray,
+  TypedArrayConstructor,
+  NumericArray,
+  NumberArray,
+  NumberArray1,
+  NumberArray2,
+  NumberArray3,
+  NumberArray6,
+  NumberArray8,
+  NumberArray9,
+  NumberArray12,
+  NumberArray16
+} from './array-types';
 export {isTypedArray, isNumericArray} from './is-array';
